Style: lyrics API 파일 생성
